+++ docs/changelog.rst
@@ -4,6 +4,12 @@
 3.x.x
 -----
 
+3.1.3
+^^^^^
+
+- |User| のユーザー名(:attr:`scapi.User.username`)が入力依存である場合の問題を修正
+- :attr:`scapi.User.lower_username`, :attr:`scapi.User.real_username` の追加
+
 3.1.2
 ^^^^^